Código desenvolvido na aula do dia 20/09/23

+++ src/main/kotlin/br/pucpr/authserver/users/UserService.kt
@@ -2,13 +2,21 @@ package br.pucpr.authserver.users
 
 import br.pucpr.authserver.exception.BadRequestException
 import br.pucpr.authserver.exception.NotFoundException
+import br.pucpr.authserver.roles.RoleRepository
+import br.pucpr.authserver.security.Jwt
+import br.pucpr.authserver.users.controller.responses.LoginResponse
+import br.pucpr.authserver.users.controller.responses.UserResponse
 import org.slf4j.LoggerFactory
 import org.springframework.data.domain.Sort
 import org.springframework.stereotype.Service
 import kotlin.jvm.optionals.getOrNull
 
 @Service
-class UserService(val repository: UserRepository) {
+class UserService(
+    val repository: UserRepository,
+    val roleRepository: RoleRepository,
+    val jwt: Jwt
+) {
     fun insert(user: User): User {
         if (repository.findByEmail(user.email) != null) {
             throw BadRequestException("User already exists")
@@ -18,7 +26,7 @@ class UserService(val repository: UserRepository) {
     }
 
     fun update(id: Long, name: String): User? {
-        val user = findByIdOrNull(id) ?: throw NotFoundException(id)
+        val user = findByIdOrThrow(id)
         if (user.name == name) return null
         user.name = name
         return repository.save(user)
@@ -29,12 +37,45 @@ class UserService(val repository: UserRepository) {
         SortDir.DESC -> repository.findAll(Sort.by("name").descending())
     }
 
+    fun findByRole(role: String): List<User> = repository.findByRole(role)
+
     fun findByIdOrNull(id: Long) = repository.findById(id).getOrNull()
+    private fun findByIdOrThrow(id: Long) =
+        findByIdOrNull(id) ?: throw NotFoundException(id)
+
     fun delete(id: Long): Boolean {
         val user = findByIdOrNull(id) ?: return false
+        if (user.roles.any { it.name == "ADMIN" }) {
+            val count = repository.findByRole("ADMIN").size
+            if (count == 1) throw BadRequestException("Cannot delete the last system admin!")
+        }
         repository.delete(user)
         log.info("User deleted: {}", user.id)
         return true
+    }
+
+    fun addRole(id: Long, roleName: String): Boolean {
+        val user = findByIdOrThrow(id)
+        if (user.roles.any { it.name == roleName }) return false
+
+        val role = roleRepository.findByName(roleName) ?:
+            throw BadRequestException("Invalid role: $roleName")
+
+        user.roles.add(role)
+        repository.save(user)
+        log.info("Granted role {} to user {}", role.name, user.id)
+        return true
+    }
+
+    fun login(email: String, password: String): LoginResponse? {
+        val user = repository.findByEmail(email) ?: return null
+        if (user.password != password) return null
+
+        log.info("User logged in. id={}, name={}", user.id, user.name)
+        return LoginResponse(
+            token = jwt.createToken(user),
+            user = UserResponse(user)
+        )
     }
 
     companion object {
